refactor(cli): move cmd/helper under cmd/internal to mirror command tree
