+++ cmd/internal/helper/docker_credentials.go
