+++ cmd/internal/helper/get_image.go
