+++ cmd/internal/helper/get_image_platforms.go
