+++ cmd/internal/helper/get_workspace_name.go
