+++ cmd/internal/helper/get_workspace_uid.go
