+++ cmd/internal/helper/ssh_git_clone.go
