+++ cmd/internal/helper/strings/strings.go
